+++ src/main/java/bitnagil/bitnagil_backend/recommendedRoutine/domain/RecommendedRoutine.java
@@ -1,8 +1,7 @@
 package bitnagil.bitnagil_backend.recommendedRoutine.domain;
 
-import bitnagil.bitnagil_backend.global.BaseTimeEntity;
+import bitnagil.bitnagil_backend.global.entity.BaseTimeEntity;
 import bitnagil.bitnagil_backend.onboarding.domain.Case;
-import bitnagil.bitnagil_backend.onboarding.domain.enums.EmotionType;
 import bitnagil.bitnagil_backend.recommendedRoutine.domain.enums.Emotion;
 import bitnagil.bitnagil_backend.recommendedRoutine.domain.enums.RecommendedRoutineLevel;
 import bitnagil.bitnagil_backend.recommendedRoutine.domain.enums.RecommendedRoutineType;
