+++ src/main/java/bitnagil/bitnagil_backend/recommendedRoutine/domain/RecommendedSubRoutine.java
@@ -1,6 +1,6 @@
 package bitnagil.bitnagil_backend.recommendedRoutine.domain;
 
-import bitnagil.bitnagil_backend.global.BaseTimeEntity;
+import bitnagil.bitnagil_backend.global.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
